+++ vite-plugin-ssr/shared/utils/projectInfo.ts
@@ -1,4 +1,4 @@
-const PROJECT_VERSION = '0.3.27'
+const PROJECT_VERSION = '0.3.28'
 
 export const projectInfo = {
   name: 'vite-plugin-ssr',